ClassServer

+++ ClassServer/ClassServer.Console/ClassWriteArg.cs
@@ -2,7 +2,7 @@ namespace ClassServer.Console;
 
 class ClassWriteArg : Any
 {
-    public virtual long Index { get; set; }
+    public virtual int Index { get; set; }
     public virtual Data Data { get; set; }
-    public virtual long MemberCount { get; set; }
+    public virtual int MemberCount { get; set; }
 }
